first procedure demo

+++ TRIOS-MOBILE/IPhoneFEPCriticalStatusViewController.m
@@ -178,7 +178,7 @@
 {
 }
 
--(int)tableView:tableView numberOfRowsInSection:(NSInteger)section
+-(long)tableView:tableView numberOfRowsInSection:(NSInteger)section
 {
    return [status.names count];
 }
